backend/requirements.txt
fastapi
uvicorn
sqlmodel
passlib==1.7.4
bcrypt==3.2.0
python-jose[cryptography]
python-multipart
python-dotenv
